Refactor loading of planet data
Rather than load and unload planet data (physics, audio, etc.) for every request when the host is not on the same planet that the request is coming from, let's load it once and only unload it if we have gone some time without a request.

I have set this timeout period to 10 seconds but perhaps another value would be preferred, or maybe it should be user configurable?