Unrefactor UpgradeEntityRequests
Or is it re-refactor?

Passing pab.DoUpgradeObject() does not work because this function is what determines whether a building can be upgraded in the first place, so with the previous changes an upgrade would succeed on the host when it should have failed due to the client that is requesting the upgrade not having the required items.