fix error when unable to obtain documents folder
for some users `Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)` returns an empty string and so Nebula will look for the player's key file within a `Dyson Sphere Program` folder inside of the game root. If this folder does not exist it will cause an error, so in this case let's create the folder first to avoid errors.