Headless now calculates planetary shields on CPU
Currently planet shields do nothing on Headless as they are emptied out and pretty much turned off as it's not possible to simulate how much energy is inside of them as energyMaxTarget is set to zero as well as isEmpty being set to True.

You can start reading more about this here: https://github.com/NebulaModTeam/nebula/issues/668#issuecomment-2227037945

While I totally understand putting 0.95 here for the headless server to assume that the shield is always attempting to be charged to 95% capacity, I believe that this is function over form.

I noticed during my testing that the RecalculatePhysicsShape method is no longer called once the variable (0.95 in this instance) became a full 1.0f, this value was read-in from the ComputeShader for PlanetAT's.

Setting it to 0.95 seemed to allow the method to be called enough so that the energyMaxTarget was updated quickly enough during the game.